small fix(es)
Small fixes. Lots of them.
Mostly rebalancing and half-dicking around the code - too lazy to write the items and abilities, so I decided to work on the interface a bit and maybe even optimise some code.
Honestly, I should rewrite everything with one array for both gonlins and dwarves but at this point I might as well develop the game into state of usability and then redo it.

I promise to fix the multiplayer issue tomorrow or something. This is the last push before I tackle it - gonna mark a major rework of the data architecture. Mostly an insane amount of Ctrl + R and quality check if everything is working and looking good.
There's gonna be blood.